Handle cases where files are already downloaded.

+++ code/downloader.py
@@ -21,22 +21,23 @@ class Downloader:
         for link in tqdm(links):
             filename = f"{collection_id}/{link[link.rfind('/')+1:]}"
             try:
-                response = self.session.get(link, stream=True)
-                response.raise_for_status()
-                with open(filename, 'wb') as fd:
-                    for chunk in response.iter_content(chunk_size=1024*1024):
-                        fd.write(chunk)
-                updated_filename = self.reproject(filename)
+                updated_filename = filename.replace('.tif', '_4326.tif')
+                if not(os.path.isfile(updated_filename)):
+                    response = self.session.get(link, stream=True)
+                    response.raise_for_status()
+                    with open(filename, 'wb') as fd:
+                        for chunk in response.iter_content(chunk_size=1024*1024):
+                            fd.write(chunk)
+                    self.reproject(filename, updated_filename)
+                    os.remove(filename)
                 downloaded_scenes.append(updated_filename)
-                os.remove(filename)
             except Exception as e:
-                print(f"Couldn't download link: {link}, Status Code: {response.status_code}, {e}")
+                print(f"Couldn't download link: {link}, {e}")
                 continue
         return downloaded_scenes
 
 
-    def reproject(self, filename):
-        updated_filename = f"{filename.replace('.tif', '_4326.tif')}"
+    def reproject(self, filename, updated_filename):
         with rasterio.open(filename) as raster:
             transform, width, height = calculate_default_transform(
                 raster.crs, DST_CRS, raster.width, raster.height, *raster.bounds
@@ -58,5 +59,5 @@ class Downloader:
                         raster_crs=raster.crs,
                         dst_transform=transform,
                         dst_crs=DST_CRS,
-                        resampling=Resampling.nearest)
-        return updated_filename
+                        resampling=Resampling.nearest
+                    )
